fix: Create a church page route fixed

+++ lib/features/church/di/church_injection_container.dart
@@ -16,22 +16,49 @@ import '../domain/usecases/update_church.dart';
 final getIt = GetIt.instance;
 
 Future<void> initChurch() async {
+  // Aguarda o Firebase Database Datasource estar pronto
+  if (!getIt.isRegistered<IFirebaseDatabaseDatasource>()) {
+    throw Exception('IFirebaseDatabaseDatasource must be registered before initializing Church feature');
+  }
+
   // Repository
-  getIt.registerLazySingleton<ChurchRepository>(
-    () => FirebaseChurchRepository(
-      firebaseDataSource: getIt<IFirebaseDatabaseDatasource>(),
-    ),
-  );
+  if (!getIt.isRegistered<ChurchRepository>()) {
+    getIt.registerLazySingleton<ChurchRepository>(
+      () => FirebaseChurchRepository(
+        firebaseDataSource: getIt<IFirebaseDatabaseDatasource>(),
+      ),
+    );
+  }
 
   // Use Cases
-  getIt.registerLazySingleton(() => CreateChurch(getIt()));
-  getIt.registerLazySingleton(() => DeleteChurch(getIt()));
-  getIt.registerLazySingleton(() => GetChurchById(getIt()));
-  getIt.registerLazySingleton(() => GetChurches(getIt()));
-  getIt.registerLazySingleton(() => GetChurchesByOwnerId(getIt()));
-  getIt.registerLazySingleton(() => UpdateChurch(getIt()));
-  getIt.registerLazySingleton(() => IncrementLikes(getIt()));
-  getIt.registerLazySingleton(() => DecrementLikes(getIt()));
-  getIt.registerLazySingleton(() => IncrementMembers(getIt()));
-  getIt.registerLazySingleton(() => DecrementMembers(getIt()));
+  if (!getIt.isRegistered<CreateChurch>()) {
+    getIt.registerLazySingleton(() => CreateChurch(getIt()));
+  }
+  if (!getIt.isRegistered<DeleteChurch>()) {
+    getIt.registerLazySingleton(() => DeleteChurch(getIt()));
+  }
+  if (!getIt.isRegistered<GetChurchById>()) {
+    getIt.registerLazySingleton(() => GetChurchById(getIt()));
+  }
+  if (!getIt.isRegistered<GetChurches>()) {
+    getIt.registerLazySingleton(() => GetChurches(getIt()));
+  }
+  if (!getIt.isRegistered<GetChurchesByOwnerId>()) {
+    getIt.registerLazySingleton(() => GetChurchesByOwnerId(getIt()));
+  }
+  if (!getIt.isRegistered<UpdateChurch>()) {
+    getIt.registerLazySingleton(() => UpdateChurch(getIt()));
+  }
+  if (!getIt.isRegistered<IncrementLikes>()) {
+    getIt.registerLazySingleton(() => IncrementLikes(getIt()));
+  }
+  if (!getIt.isRegistered<DecrementLikes>()) {
+    getIt.registerLazySingleton(() => DecrementLikes(getIt()));
+  }
+  if (!getIt.isRegistered<IncrementMembers>()) {
+    getIt.registerLazySingleton(() => IncrementMembers(getIt()));
+  }
+  if (!getIt.isRegistered<DecrementMembers>()) {
+    getIt.registerLazySingleton(() => DecrementMembers(getIt()));
+  }
 }
